app/readinglist2md/requirements.txt
aws-lambda-powertools==3.2.0
boto3==1.35.57
Jinja2==3.1.4
notion-client==2.2.1
pandas==2.2.3
python-dotenv==1.0.1